# docs/swagger-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/arc-api/v1/projects/{projectId}/subsystems/get": {
      "post": {
        "operationId": "SubsystemController_getSubsystems",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the project",
            "schema": {
              "example": "12345",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "List of subsystem system ids",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemIdsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubsystemDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Some subsystems are not found"
          },
          "422": {
            "description": "Failed to get subsystems"
          }
        },
        "summary": "Get subsystems for provided systemIds",
        "tags": [
          "subsystems"
        ]
      }
    },
    "/arc-api/v1/projects/{projectId}/subsystems/{subsystemSystemId}/components/get": {
      "post": {
        "operationId": "SubsystemController_getComponentsInSubsystem",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the project",
            "schema": {
              "example": "12345",
              "type": "string"
            }
          },
          {
            "name": "subsystemSystemId",
            "required": true,
            "in": "path",
            "description": "Subsystem system ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "description": "List of system ids for usecases. Optional.\n\nIf provided, only top level components in the subystem for these usecases will be returned.\n\nOtherwise, all top level of components in the subsystem will be returned.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemIdsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BaseComponentDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Subsystem cannot be found"
          },
          "422": {
            "description": "Failed to get components in subsystem)"
          }
        },
        "summary": "Get all top level components in a subsystem for provided usecases",
        "tags": [
          "subsystems"
        ]
      }
    },
...
  },
  "info": {
    "title": "AudioReach Creator API",
    "description": "Audio Reach Creator API is an open source and cross platform backend framework for updating Audio Reach database files.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "EndPointLink": {
        "type": "object",
        "properties": {
          "hypertextRef": {
            "type": "string",
            "description": "Hypertext reference URL following 'project{projectId}'"
          },
          "method": {
            "type": "string",
            "description": "HTTP method",
            "example": "GET"
          },
          "description": {
            "type": "string",
            "description": "Description of the endpoint"
          }
        },
        "required": [
          "hypertextRef",
          "method",
          "description"
        ]
      },
...
      "BaseComponentDto": {
        "type": "object",
        "properties": {
          "systemId": {
            "type": "string",
            "description": "System ID"
          },
          "changeId": {
            "type": "string",
            "description": "Change ID"
          },
          "editType": {
            "type": "string",
            "description": "Edit type",
            "enum": [
              "Added",
              "Updated",
              "Deleted",
              "None"
            ]
          },
          "id": {
            "type": "object",
            "description": "Component ID"
          },
          "name": {
            "type": "string",
            "description": "Component name"
          },
          "componentType": {
            "type": "string",
            "description": "Component type",
            "enum": [
              "Module",
              "Subsystem",
              "Container",
              "DataLink",
              "ControlLink",
              "Subgraph"
            ]
          },
          "relatedEndPointLinks": {
            "description": "Related endpoint links",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EndPointLink"
            }
          }
        },
        "required": [
          "systemId",
          "changeId",
          "editType",
          "id",
          "name",
          "componentType",
          "relatedEndPointLinks"
        ]
      },
...
      "DataPortDto": {
        "type": "object",
        "properties": {
          "systemId": {
            "type": "string",
            "description": "System ID"
          },
          "changeId": {
            "type": "string",
            "description": "Change ID"
          },
          "editType": {
            "type": "string",
            "description": "Edit type",
            "enum": [
              "Added",
              "Updated",
              "Deleted",
              "None"
            ]
          },
          "id": {
            "type": "object",
            "description": "Component ID"
          },
          "name": {
            "type": "string",
            "description": "Component name"
          },
          "componentType": {
            "type": "string",
            "description": "Component type",
            "enum": [
              "Module",
              "Subsystem",
              "Container",
              "DataLink",
              "ControlLink",
              "Subgraph"
            ]
          },
          "relatedEndPointLinks": {
            "description": "Related endpoint links",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EndPointLink"
            }
          },
          "portIoType": {
            "type": "string",
            "description": "Port IO type",
            "enum": [
              "Input",
              "Output"
            ]
          },
          "portType": {
            "type": "string",
            "description": "Port type",
            "enum": [
              "Static",
              "Dynamic"
            ]
          }
        },
        "required": [
          "systemId",
          "changeId",
          "editType",
          "id",
          "name",
          "componentType",
          "relatedEndPointLinks",
          "portIoType",
          "portType"
        ]
      },
      "ControlPortIntentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Intent ID"
          },
          "name": {
            "type": "string",
            "description": "Intent name"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "ControlPortDto": {
        "type": "object",
        "properties": {
          "systemId": {
            "type": "string",
            "description": "System ID"
          },
          "changeId": {
            "type": "string",
            "description": "Change ID"
          },
          "editType": {
            "type": "string",
            "description": "Edit type",
            "enum": [
              "Added",
              "Updated",
              "Deleted",
              "None"
            ]
          },
          "id": {
            "type": "object",
            "description": "Component ID"
          },
          "name": {
            "type": "string",
            "description": "Component name"
          },
          "componentType": {
            "type": "string",
            "description": "Component type",
            "enum": [
              "Module",
              "Subsystem",
              "Container",
              "DataLink",
              "ControlLink",
              "Subgraph"
            ]
          },
          "relatedEndPointLinks": {
            "description": "Related endpoint links",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EndPointLink"
            }
          },
          "portType": {
            "type": "string",
            "description": "Port type",
            "enum": [
              "Static",
              "Dynamic"
            ]
          },
          "controlPortName": {
            "type": "string",
            "description": "Control port name"
          },
          "intents": {
            "description": "Control port intents",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ControlPortIntentDto"
            }
          }
        },
        "required": [
          "systemId",
          "changeId",
          "editType",
          "id",
          "name",
          "componentType",
          "relatedEndPointLinks",
          "portType",
          "controlPortName",
          "intents"
        ]
      },
...
      "SystemIdsRequestDto": {
        "type": "object",
        "properties": {
          "systemIds": {
            "description": "Array of system IDs",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "systemIds"
        ]
      },
      "KeyInfo": {
        "type": "object",
        "properties": {
          "keyId": {
            "type": "number",
            "description": "Key id"
          },
          "keyLabel": {
            "type": "string",
            "description": "Key name"
          }
        },
        "required": [
          "keyId",
          "keyLabel"
        ]
      },
      "SubsystemDto": {
        "type": "object",
        "properties": {
          "systemId": {
            "type": "string",
            "description": "System ID"
          },
          "changeId": {
            "type": "string",
            "description": "Change ID"
          },
          "editType": {
            "type": "string",
            "description": "Edit type",
            "enum": [
              "Added",
              "Updated",
              "Deleted",
              "None"
            ]
          },
          "id": {
            "type": "object",
            "description": "Component ID"
          },
          "name": {
            "type": "string",
            "description": "Component name"
          },
          "componentType": {
            "type": "string",
            "description": "Component type",
            "enum": [
              "Module",
              "Subsystem",
              "Container",
              "DataLink",
              "ControlLink",
              "Subgraph"
            ]
          },
          "relatedEndPointLinks": {
            "description": "Related endpoint links",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EndPointLink"
            }
          },
          "parentId": {
            "type": "object",
            "description": "Parent component ID"
          },
          "dataPorts": {
            "description": "Data ports",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataPortDto"
            }
          },
          "controlPorts": {
            "description": "Control ports",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ControlPortDto"
            }
          },
          "_filteredKeys": {
            "description": "Filtered keys assigned to the subsystem",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KeyInfo"
            }
          }
        },
        "required": [
          "systemId",
          "changeId",
          "editType",
          "id",
          "name",
          "componentType",
          "relatedEndPointLinks",
          "dataPorts",
          "controlPorts",
          "_filteredKeys"
        ]
      },
...
    }
  }
}